+++ src/main/java/com/zen/fogman/client/renderers/ModRenderers.java
@@ -9,6 +9,7 @@ public class ModRenderers {
     public static void registerRenderers() {
         EntityRendererRegistry.register(ModEntities.THE_MAN, TheManRenderer::new);
         EntityRendererRegistry.register(ModEntities.THE_MAN_HALLUCINATION, TheManRenderer::new);
+        EntityRendererRegistry.register(ModEntities.THE_MAN_PARANOIA, TheManRenderer::new);
         EntityRendererRegistry.register(ModEntities.THE_MAN_SPIT, TheManSpitRenderer::new);
     }
 }