+++ src/main/java/com/zen/fogman/common/entity/the_man/TheManPredicates.java
@@ -32,7 +32,7 @@ public class TheManPredicates {
             entity instanceof TheManEntity theMan && !theMan.isHallucination() && theMan.isAlive();
 
     public static final Predicate<Entity> VALID_MAN_HALLUCINATION = entity ->
-            entity instanceof TheManEntity theMan && theMan.isHallucination() && theMan.isAlive();
+            entity instanceof TheManEntity theMan && theMan.isHallucination() && !theMan.isParanoia() && theMan.isAlive();
 
     public static final Predicate<BlockState> EXCEPT_AIR = blockState -> !blockState.isAir();
 }